demo/requirements.txt
torch==2.4.1
transformers==4.45
accelerate==0.34.2
spacy==3.7.4
numpy==2.0.2
openai==1.51.2
jieba==0.42.1
chardet==5.2.0
pysbd==0.3.4
pypdfium2==4.30.1
python-docx==1.1.2
streamlit==1.40.1